+++ src/Input.test.jsx
@@ -1,0 +1,31 @@
+import { render, fireEvent } from '@testing-library/react';
+
+import Input from './Input';
+
+import { TASK_TITLE } from './stubs';
+
+describe('Input', () => {
+  const onChangeTitle = jest.fn();
+  const onClickAddTask = jest.fn();
+
+  const renderInputComponent = () => render(
+    <Input value={TASK_TITLE} onChange={onChangeTitle} onClick={onClickAddTask} />,
+  );
+  it('투두 인풋의 버튼을 클릭하면 onClickAddTask 이벤트가 호출된다', () => {
+    const { getByText } = renderInputComponent();
+
+    const addTodoButton = getByText('추가');
+    addTodoButton.click();
+
+    expect(onClickAddTask).toBeCalled();
+  });
+
+  it('투두 인풋에 타이핑을 하면 onChangeTitle 이벤트가 호출된다', () => {
+    const { getByPlaceholderText } = renderInputComponent();
+
+    const toDoInput = getByPlaceholderText('할 일을 입력해 주세요');
+    fireEvent.change(toDoInput, { target: { value: '타이핑을 해주세요' } });
+
+    expect(onChangeTitle).toBeCalled();
+  });
+});
